feat: migrate to default Python Method
