+++ multiauth/utils.py
@@ -198,14 +198,14 @@ def uncurl(curl: str) -> ParsedCurlContent:
 
     # Now we have to analyze what we have
     # First find out what type of method are we using
-    method: HTTPMethod = 'GET'
+    method: HTTPMethod = HTTPMethod.GET
     if parsed_args.request:
-        method = parsed_args.request.upper()
+        method = HTTPMethod(parsed_args.request.upper())
     else:
         if parsed_args.data:
-            method = 'POST'
+            method = HTTPMethod.POST
         else:
-            method = 'GET'
+            method = HTTPMethod.GET
 
     # Now we have to extract the headers
     headers: dict[str, str] = {}
